You are given two arrays A and B consisting of N Integers each.
Index K is named fair if the four sums (AO)++AD-1D) (ADG++AIN-1D).
(BIC++ BIK-1) and (BIK-BIN-1) are all equal. In other words, K is the
index where the two arrays, A and B, can be

split (into two non-empty arrays each) in such a way that the sums of
the resulting arrays' elements are equal.

For example, given arrays A-10,4-1,0, 31 and B-10-2,5, 0, 31, Index K-3
is fair. The sums of the subarrays are all equal:
0+4+(-1)-3,0+3+30+2)+5-3 and 0+3-3. On the other hand, index K-2 is not
fair the sums of the subarrays are: 0+4-4 (-1)+0+3-20+(-2) -2 and
5+0+3-8.

Signed-off-by: rohi92 <gorthyrohith@gmail.com>